# cmd/api-server/docs/swagger.yaml
basePath: /
definitions:
  handler.CreateURLRequest:
    properties:
      original_url:
        type: string
    required:
    - original_url
    type: object
  handler.CreateURLResponse:
    properties:
      original_url:
        type: string
      short_url:
        type: string
    type: object
  handler.createUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  handler.userLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.userLoginResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
  utils.APIError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  utils.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/utils.APIError'
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for shortening URLs
  title: URL Shortener API
  version: "1.0"
paths:
  /{shortUrl}:
    get:
      description: Looks up the short code and redirects the client to the original
        URL
      responses:
        "400":
          description: url code is required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: short url not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Redirect to original URL
      tags:
      - urls
  /api/urls/shorten:
    post:
      consumes:
      - application/json
      description: Accepts a long URL and returns a shortened version
      parameters:
      - description: URL to shorten
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateURLResponse'
        "400":
          description: invalid request payload / validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Shorten a URL
      tags:
      - urls
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns an access token and refresh token
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.userLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userLoginResponse'
        "400":
          description: invalid request payload
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: invalid credentials
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/utils.Envelope'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/utils.Envelope'
      summary: Login user
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.userLoginResponse'
        "400":
          description: email and password required / invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: email already taken
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: something went wrong
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"